% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_server.R
\name{check_fitbit_notification_signature}
\alias{check_fitbit_notification_signature}
\title{Check subscriber signature
https://dev.fitbit.com/build/reference/web-api/developer-guide/best-practices/#Subscriber-Security}
\usage{
check_fitbit_notification_signature(req, res)
}
\arguments{
\item{req}{Request object}
}
\description{
Check subscriber signature
https://dev.fitbit.com/build/reference/web-api/developer-guide/best-practices/#Subscriber-Security
}
\keyword{internal}
