% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_auth.R
\name{create_fitbit_auth_state}
\alias{create_fitbit_auth_state}
\title{Fitbit authorization URL with PKCE}
\usage{
create_fitbit_auth_state(uid)
}
\arguments{
\item{uid}{User ID.}
}
\value{
Fitbit authorization URL.
}
\description{
NOTE: With local memory cache, this is works only in a single R session.
}
