% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_auth.R
\name{FitbitToken}
\alias{FitbitToken}
\title{Fitbit token}
\description{
Fitbit token

Fitbit token
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token_type}}{Type}

\item{\code{.access_token}}{Access token}

\item{\code{expires_at}}{Expiration time}

\item{\code{refresh_token}}{Refresh token}

\item{\code{scope}}{Scope}

\item{\code{user_id}}{Fitbit user ID}

\item{\code{uid}}{Firebase user ID}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{access_token}}{Returns the access token, refreshing if necessary.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FitbitToken-new}{\code{FitbitToken$new()}}
\item \href{#method-FitbitToken-is_expired}{\code{FitbitToken$is_expired()}}
\item \href{#method-FitbitToken-refresh}{\code{FitbitToken$refresh()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitToken-new}{}}}
\subsection{Method \code{new()}}{
Initialize Fitbit token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitToken$new(code = NULL, pkce = NULL, uid = NULL, test = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{Authorization code.}

\item{\code{pkce}}{PKCE.}

\item{\code{uid}}{Firebase user ID.}

\item{\code{test}}{Test mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitToken-is_expired"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitToken-is_expired}{}}}
\subsection{Method \code{is_expired()}}{
Check if token is expired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitToken$is_expired()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitToken-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitToken-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitToken$refresh(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{Force refresh.}
}
\if{html}{\out{</div>}}
}
}
}
