% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitbitWebAPI.R
\name{FitbitWebAPI}
\alias{FitbitWebAPI}
\title{Fitbit Web API endpoints}
\description{
Fitbit Web API endpoints

Fitbit Web API endpoints
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{A Fitbit user token}

\item{\code{rate_limit_key}}{The rate limit key unique to the user}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FitbitWebAPI-new}{\code{FitbitWebAPI$new()}}
\item \href{#method-FitbitWebAPI-query}{\code{FitbitWebAPI$query()}}
\item \href{#method-FitbitWebAPI-resp_store_rate_limit}{\code{FitbitWebAPI$resp_store_rate_limit()}}
\item \href{#method-FitbitWebAPI-limit_rate}{\code{FitbitWebAPI$limit_rate()}}
\item \href{#method-FitbitWebAPI-req_fitbit_error}{\code{FitbitWebAPI$req_fitbit_error()}}
\item \href{#method-FitbitWebAPI-body_log_weight}{\code{FitbitWebAPI$body_log_weight()}}
\item \href{#method-FitbitWebAPI-nutrition_time_series}{\code{FitbitWebAPI$nutrition_time_series()}}
\item \href{#method-FitbitWebAPI-activity_time_series}{\code{FitbitWebAPI$activity_time_series()}}
\item \href{#method-FitbitWebAPI-activity_intraday}{\code{FitbitWebAPI$activity_intraday()}}
\item \href{#method-FitbitWebAPI-create_subscription}{\code{FitbitWebAPI$create_subscription()}}
\item \href{#method-FitbitWebAPI-delete_subscription}{\code{FitbitWebAPI$delete_subscription()}}
\item \href{#method-FitbitWebAPI-list_subscriptions}{\code{FitbitWebAPI$list_subscriptions()}}
\item \href{#method-FitbitWebAPI-clone}{\code{FitbitWebAPI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-new"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-new}{}}}
\subsection{Method \code{new()}}{
Initialize the FitbitWebAPI object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$new(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{A Fitbit user token}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-query"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-query}{}}}
\subsection{Method \code{query()}}{
Query the Fitbit Web API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$query(ep, method = "GET", error_message = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ep}}{character vector of the endpoint}

\item{\code{method}}{The HTTP method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-resp_store_rate_limit"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-resp_store_rate_limit}{}}}
\subsection{Method \code{resp_store_rate_limit()}}{
Store the rate limit information
Currently using a app wide rate limiter cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$resp_store_rate_limit(resp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resp}}{The response object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-limit_rate"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-limit_rate}{}}}
\subsection{Method \code{limit_rate()}}{
Limit the rate of API calls
Now the response is just to wait, which is not ideal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$limit_rate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-req_fitbit_error"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-req_fitbit_error}{}}}
\subsection{Method \code{req_fitbit_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$req_fitbit_error(req, error_message = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-body_log_weight"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-body_log_weight}{}}}
\subsection{Method \code{body_log_weight()}}{
Get the weight log
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$body_log_weight(
  start_date,
  end_date = NULL,
  period = c("1d", "7d", "30d", "1w", "1m")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The start date}

\item{\code{end_date}}{The end date. Use either end_date or period. Range can be at most 31 days.}

\item{\code{period}}{The period of the log. One of "1d", "7d", "30d", "1w", "1m".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-nutrition_time_series"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-nutrition_time_series}{}}}
\subsection{Method \code{nutrition_time_series()}}{
Get nutrition time series
https://dev.fitbit.com/build/reference/web-api/nutrition-timeseries/get-nutrition-timeseries-by-date-range/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$nutrition_time_series(
  start_date,
  end_date = NULL,
  period = c("1d", "7d", "30d", "1w", "1m", "3m", "6m", "1y"),
  resource = c("caloriesIn", "water")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The start date}

\item{\code{end_date}}{The end date. Use either end_date or period. Range can be at most 1095 days.}

\item{\code{period}}{The period of the log. One of "1d", "7d", "30d", "1w", "1m", "3m", "6m", "1y".}

\item{\code{resource}}{The resource to query. One of "caloriesIn", "water".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-activity_time_series"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-activity_time_series}{}}}
\subsection{Method \code{activity_time_series()}}{
Get activity time series
https://dev.fitbit.com/build/reference/web-api/intraday/get-activity-intraday-by-date/
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$activity_time_series(
  start_date,
  end_date = NULL,
  period = c("1d", "7d", "30d", "1w", "1m", "3m", "6m", "1y"),
  resource = c("calories", "activityCalories", "caloriesBMR", "distance", "elevation",
    "floors", "minutesSedentary", "minutesLightlyActive", "minutesFairlyActive",
    "minutesVeryActive", "steps", "swimming-strokes")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The start date}

\item{\code{end_date}}{The end date. Use either end_date or period. Range can be at most 1095 days, except for activityCalories which is 30 days.}

\item{\code{period}}{The period of the log. One of "1d", "7d", "30d", "1w", "1m", "3m", "6m", "1y".}

\item{\code{resource}}{The resource to query. One of "calories", "activityCalories", "caloriesBMR", "distance", "elevation", "floors", "minutesSedentary", "minutesLightlyActive", "minutesFairlyActive", "minutesVeryActive", "steps", "swimming-strokes".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-activity_intraday"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-activity_intraday}{}}}
\subsection{Method \code{activity_intraday()}}{
Get the intraday data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$activity_intraday(
  date = Sys.Date(),
  resource = c("calories", "distance", "elevation", "floors", "steps",
    "swimming-strokes"),
  detail_level = c("15min", "5min", "1min")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{The date}

\item{\code{resource}}{The resource to query. One of "calories", "distance", "elevation", "floors", "steps", "swimming-strokes".}

\item{\code{detail_level}}{The detail level. One of "1min", "5min", "15min". Default is "15min".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-create_subscription"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-create_subscription}{}}}
\subsection{Method \code{create_subscription()}}{
Create a subscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$create_subscription(
  collection_path = c("activities", "body", "foods", "sleep", "userRevokedAccess", "all"),
  subscription_id
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_path}}{The collection path. One of "activities", "body", "foods", "sleep", "userRevokedAccess", "all".}

\item{\code{subscription_id}}{The subscription id. Max length 50 characters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-delete_subscription"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-delete_subscription}{}}}
\subsection{Method \code{delete_subscription()}}{
Delete a subscription
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$delete_subscription(
  collection_path = c("activities", "body", "foods", "sleep", "userRevokedAccess", "all"),
  subscription_id
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_path}}{The collection path. One of "activities", "body", "foods", "sleep", "userRevokedAccess", "all".}

\item{\code{subscription_id}}{The subscription id. Max length 50 characters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-list_subscriptions"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-list_subscriptions}{}}}
\subsection{Method \code{list_subscriptions()}}{
List subscriptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$list_subscriptions(
  collection_path = c("activities", "body", "foods", "sleep", "userRevokedAccess", "all")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection_path}}{The collection path. One of "activities", "body", "foods", "sleep", "userRevokedAccess", "all".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FitbitWebAPI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FitbitWebAPI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitbitWebAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
