% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_server.R
\name{handle_fibit_subscriber_verification}
\alias{handle_fibit_subscriber_verification}
\title{Answer Fitbit subscriber verification request}
\usage{
handle_fibit_subscriber_verification(req, res)
}
\arguments{
\item{req}{Request object}
}
\description{
Answer Fitbit subscriber verification request
}
\keyword{internal}
