% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_server.R
\name{handle_fitbit_notifications}
\alias{handle_fitbit_notifications}
\title{Handle Fitbit notification}
\usage{
handle_fitbit_notifications(req)
}
\arguments{
\item{req}{Request object}
}
\description{
Handle Fitbit notification
}
\keyword{internal}
