% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_auth.R
\name{make_oauth_url}
\alias{make_oauth_url}
\title{Fitbit authorization URL}
\usage{
make_oauth_url(pkce, state)
}
\arguments{
\item{pkce}{PKCE.}

\item{state}{User specific state sent to authorization.}
}
\value{
Fitbit authorization URL.
}
\description{
Fitbit authorization URL
}
