% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_auth.R
\name{refresh_fitbit_access_token}
\alias{refresh_fitbit_access_token}
\title{Refresh token, not working}
\usage{
refresh_fitbit_access_token(token)
}
\arguments{
\item{token}{Token.}
}
\description{
Refresh token, not working
}
