% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_server.R
\name{refresh_fitbit_data}
\alias{refresh_fitbit_data}
\title{Refresh fitbit data history}
\usage{
refresh_fitbit_data(uid)
}
\arguments{
\item{uid}{User ID}
}
\description{
Refresh fitbit data history
}
\keyword{internal}
