% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_auth.R
\name{request_fitbit_access_token}
\alias{request_fitbit_access_token}
\title{Get token from Fitbit}
\usage{
request_fitbit_access_token(code, pkce)
}
\arguments{
\item{code}{Authorization code.}

\item{pkce}{PKCE.}
}
\value{
Token.
}
\description{
Get token from Fitbit
}
